move meta title & meta setting generator to  ~lib/meta
