+++ pages/provinces/[provinceSlug]/index.tsx
@@ -6,6 +6,7 @@ import { PageContent } from "~/components/layout/page-content";
 import { PageHeader } from "~/components/layout/page-header";
 import { SearchForm } from "~/components/search-form";
 import { useSearch } from "~/lib/hooks/use-search";
+import { getProvinceMeta } from "~/lib/meta";
 import provinces, { Contact, getProvincesPaths } from "~/lib/provinces";
 import { getTheLastSegmentFromKebabCase } from "~/lib/string-utils";
 
@@ -17,18 +18,6 @@ type ProvinceProps = {
   provinceName: string;
   provinceSlug: string;
   contactList: Contact[];
-};
-
-export function getProvinceMetaTitle(provinceName: string) {
-  return `Informasi Faskes & Alkes untuk COVID-19 di Provinsi ${provinceName}`;
-}
-
-const getMeta = (provinceName: string) => {
-  return {
-    // @TODO: change this after got a better title
-    title: getProvinceMetaTitle(provinceName),
-    description: `Informasi seputar COVID-19 dan kontak fasilitas/alat kesehatan di Provinsi ${provinceName} yang dikumpulkan relawan melalui pencarian di internet atau media sosial.`,
-  };
 };
 
 export default function ProvincePage(props: ProvinceProps) {
@@ -72,7 +61,7 @@ export default function ProvincePage(props: ProvinceProps) {
 
   // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
   if (provinceName) {
-    const meta = getMeta(provinceName);
+    const meta = getProvinceMeta(provinceName);
 
     return (
       <Page>
